Another minor nit: Make sure the constant here is a float so the compiler doesn't promote the entire expression to double.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@216248 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ lib/msun/src/e_log2f.c
@@ -50,7 +50,7 @@ __ieee754_log2f(float x)
 	SET_FLOAT_WORD(x,hx|(i^0x3f800000));	/* normalize x or x/2 */
 	k += (i>>23);
 	f = __kernel_logf(x);
-	x = x - 1;
+	x = x - (float)1.0;
 	GET_FLOAT_WORD(hx,x);
 	SET_FLOAT_WORD(hi,hx&0xfffff000);
 	lo = x - hi;
